+++ regimes/regimes.go
@@ -29,4 +29,5 @@ import (
 	_ "github.com/invopop/gobl/regimes/se"
 	_ "github.com/invopop/gobl/regimes/sg"
 	_ "github.com/invopop/gobl/regimes/us"
+	_ "github.com/invopop/gobl/regimes/uy"
 )